restructure file tree for clarity
