--- shells/pipes-shell-2/planner.js
@@ -1,13 +1,0 @@
-/**
- * @license
- * Copyright 2019 Google LLC.
- * This code may only be used under the BSD style license found at
- * http://polymer.github.io/LICENSE.txt
- * Code distributed by Google as part of this project is also
- * subject to an additional IP rights grant found at
- * http://polymer.github.io/PATENTS.txt
- */
-export const initPlanner = (context) => {
-  const names = context.allRecipes.map(r => r.name);
-  console.warn(names);
-};
