+++ shells/pipes-shell-2/source/bus.js
@@ -8,8 +8,7 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-import {logsFactory} from '../../build/runtime/log-factory.js';
-
+import {logsFactory} from '../../../build/runtime/log-factory.js';
 const {log} = logsFactory('BUS', '#d32e1b');
 
 export const Bus = class {
