+++ shells/pipes-shell-2/source/dispatcher.js
