+++ shells/pipes-shell-2/source/pipes-api.js
@@ -9,9 +9,9 @@
  */
 
 import {generateId} from '../../../modalities/dom/components/generate-id.js';
-import {storeByTag} from '../lib/utils.js';
-import {requirePipesArc} from '../lib/pipes-arc.js';
-import {initPipeStore, mirrorStore} from '../context.js';
+import {storeByTag} from './lib/utils.js';
+import {requirePipesArc} from './lib/pipes-arc.js';
+import {initPipeStore, mirrorStore} from './context.js';
 
 let pipes;
 let pipeStore;
