+++ shells/pipes-shell-2/source/schemas.js
