+++ shells/pipes-shell-2/source/smoke.js
