+++ shells/pipes-shell-2/source/verbs/caption.js
